requirements.txt
requests_oauthlib
streamlit
dotenv
google-generativeai